@@ -71,36 +71,30 @@
         "patreon": "^0.4.1",
         "pluralize": "^7.0.0",
         "pmx": "^1.6.7",
-        "pug": "^2.0.4",
-        "pug-loader": "^2.4.0",
         "react-test-renderer": "^16.8.6",
         "redis": "^3.1.2",
         "redux-loop": "^4.5.2",
         "regenerator-runtime": "^0.13.3",
         "sass": "^1.74.1",
-        "socket.io": "^2.4.1",
+        "socket.io": "^4.7.5",
         "underscore": "^1.13.1",
         "uuid": "^3.1.0",
-        "webpack-cli": "^3.3.6",
-        "webpack-dev-middleware": "^3.7.0",
+        "vite": "^5.2.11",
+        "vite-express": "^0.16.0",
         "winston": "^3.11.0",
         "winston-daily-rotate-file": "^4.7.1"
     },
     "devDependencies": {
+        "@rollup/plugin-inject": "^5.0.5",
+        "@vitejs/plugin-react-swc": "^3.6.0",
         "babel-eslint": "^8.2.3",
-        "babel-loader": "^8.0.5",
         "classnames": "^2.2.5",
-        "css-loader": "^3.1.0",
-        "eslint": "^4.19.1",
-        "eslint-plugin-react": "^7.8.2",
-        "event-source-polyfill": "0.0.12",
-        "file-loader": "^1.1.11",
+        "eslint": "^8.57.0",
+        "eslint-plugin-react": "^7.34.1",
         "jquery": "^3.4.1",
-        "jquery-migrate": "=3.0.0",
-        "jquery-validation": "^1.17.0",
+        "jquery-validation": "^1.20.0",
         "jquery-validation-unobtrusive": "^3.2.11",
-        "less": "^3.9.0",
-        "less-loader": "^4.1.0",
+        "less": "^4.2.0",
         "moment": "^2.22.1",
         "node-noop": "^1.0.0",
         "path-to-regexp": "^2.2.1",
