@@ -1,9 +1,8 @@
-const moment = require('moment');
-
-const AgendaRules = require('./AgendaRules');
-const DeckWrapper = require('./DeckWrapper');
-const Formats = require('./Formats');
-const RestrictedList = require('./RestrictedList');
+import moment from 'moment';
+import AgendaRules from './AgendaRules.js';
+import DeckWrapper from './DeckWrapper.js';
+import Formats from './Formats/index.js';
+import RestrictedList from './RestrictedList.js';
 
 class DeckValidator {
     constructor(packs, restrictedListRules, customRules = {}) {
@@ -120,4 +119,4 @@ class DeckValidator {
     }
 }
 
-module.exports = DeckValidator;
+export default DeckValidator;
