@@ -1,7 +1,6 @@
-const passport = require('passport');
-
-const { wrapAsync } = require('../util');
-const ServiceFactory = require('../services/ServiceFactory.js');
+import passport from 'passport';
+import { wrapAsync } from '../util.js';
+import ServiceFactory from '../services/ServiceFactory.js';
 
 function extractEventFromRequest(req) {
     const {
@@ -37,7 +36,7 @@ function extractEventFromRequest(req) {
     };
 }
 
-module.exports.init = function(server, options) {
+export const init = function(server, options) {
     const eventService = ServiceFactory.eventService(options.db);
 
     server.get('/api/events', wrapAsync(async function(req, res) {
