@@ -1,4 +1,4 @@
-module.exports = [
+export default [
     { name: 'House Baratheon', value: 'baratheon' },
     { name: 'House Greyjoy', value: 'greyjoy' },
     { name: 'House Lannister', value: 'lannister' },
