@@ -1,22 +1,19 @@
-const express = require('express');
+import express from 'express';
+import ViteExpress from 'vite-express';
 const app = express();
-const bodyParser = require('body-parser');
-const passport = require('passport');
-const logger = require('./log.js');
-const api = require('./api');
-const path = require('path');
-const http = require('http');
-const webpackDevMiddleware = require('webpack-dev-middleware');
-const webpackHotMiddleware = require('webpack-hot-middleware');
-const historyApiFallback = require('connect-history-api-fallback');
-const webpack = require('webpack');
-const webpackConfig = require('../webpack.dev.js');
-const Sentry = require('@sentry/node');
-const passportJwt = require('passport-jwt');
+import bodyParser from 'body-parser';
+import passport from 'passport';
+import logger from './log.js';
+import {init as ApiInit} from './api/index.js';
+import http from 'http';
+import Sentry from '@sentry/node';
+import passportJwt from 'passport-jwt';
 const JwtStrategy = passportJwt.Strategy;
 const ExtractJwt = passportJwt.ExtractJwt;
 
-const ServiceFactory = require('./services/ServiceFactory.js');
+import ServiceFactory from './services/ServiceFactory.js';
+
+const __dirname = import.meta.dirname;
 
 class Server {
     constructor(isDeveloping) {
@@ -63,46 +60,14 @@ class Server {
         app.use(bodyParser.json());
         app.use(bodyParser.urlencoded({ extended: false }));
 
-        api.init(app, options);
+        ApiInit(app, options);
 
         app.use(express.static(__dirname + '/../public'));
         app.use(express.static(__dirname + '/../dist'));
 
-        if (this.isDeveloping) {
-            const compiler = webpack(webpackConfig);
-            const middleware = webpackDevMiddleware(compiler, {
-                hot: true,
-                contentBase: 'client',
-                publicPath: '/',
-                stats: {
-                    colors: true,
-                    hash: false,
-                    timings: true,
-                    chunks: false,
-                    chunkModules: false,
-                    modules: false
-                },
-                historyApiFallback: true
-            });
-
-            app.set('view engine', 'pug');
-            app.set('views', path.join(__dirname, '..', 'views'));
-
-            app.use(middleware);
-            app.use(
-                webpackHotMiddleware(compiler, {
-                    log: false,
-                    path: '/__webpack_hmr',
-                    heartbeat: 2000
-                })
-            );
-            app.use(historyApiFallback());
-            app.use(middleware);
-        } else {
-            app.get('*', (req, res) => {
-                res.sendFile(path.join(__dirname, '..', 'dist', 'index.html'));
-            });
-        }
+        // app.get('*', (req, res) => {
+        //     res.sendFile(path.join(__dirname, '..', 'dist', 'index.html'));
+        // });
 
         // Define error middleware last
         app.use(function (err, req, res, next) {
@@ -120,6 +85,8 @@ class Server {
 
     run() {
         let port = process.env.PORT || this.configService.getValue('port') || 4000;
+
+        ViteExpress.bind(app, this.server);
 
         this.server.listen(port, '0.0.0.0', function onStart(err) {
             if (err) {
@@ -141,4 +108,4 @@ class Server {
     }
 }
 
-module.exports = Server;
+export default Server;
