@@ -1,9 +1,8 @@
-const logger = require('./log.js');
-const EventEmitter = require('events');
-const jwt = require('jsonwebtoken');
-const Sentry = require('@sentry/node');
-
-const User = require('./models/User');
+import logger from './log.js';
+import EventEmitter from 'events';
+import jwt from 'jsonwebtoken';
+import Sentry from '@sentry/node';
+import User from './models/User.js';
 
 class Socket extends EventEmitter {
     constructor(socket, options = {}) {
@@ -81,4 +80,4 @@ class Socket extends EventEmitter {
     }
 }
 
-module.exports = Socket;
+export default Socket;
