@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { validateDeck } from '../../../deck-helper';
+import {validateDeck} from '../../../deck-helper';
 
 import Input from '../Form/Input';
 import Select from '../Form/Select';
@@ -33,7 +33,7 @@ class DeckEditor extends React.Component {
             eventId: undefined
         };
 
-        if(props.deck) {
+        if (props.deck) {
             this.state.deckId = props.deck._id;
             this.state.deckName = props.deck.name;
             this.state.plotCards = props.deck.plotCards;
@@ -46,11 +46,11 @@ class DeckEditor extends React.Component {
             this.state.eventId = props.deck.eventId;
 
             let cardList = '';
-            for(const card of props.deck.drawCards) {
+            for (const card of props.deck.drawCards) {
                 cardList += this.formatCardListItem(card) + '\n';
             }
 
-            for(const plot of props.deck.plotCards) {
+            for (const plot of props.deck.plotCards) {
                 cardList += this.formatCardListItem(plot) + '\n';
             }
 
@@ -63,7 +63,7 @@ class DeckEditor extends React.Component {
     }
 
     componentWillReceiveProps(props) {
-        if(props.factions && !this.state.faction) {
+        if (props.factions && !this.state.faction) {
             this.setState({ faction: props.factions['baratheon'] }, this.triggerDeckUpdated);
         }
     }
@@ -80,11 +80,13 @@ class DeckEditor extends React.Component {
             eventId: this.state.eventId
         };
 
-        if(!this.props.restrictedList && !this.props.currentRestrictedList) {
+        if (!this.props.restrictedList && !this.props.currentRestrictedList) {
             deck.status = {};
         } else {
             const selectedRestrictedList = restrictedList || this.props.currentRestrictedList;
-            const restrictedLists = selectedRestrictedList ? [selectedRestrictedList] : this.props.restrictedList;
+            const restrictedLists = selectedRestrictedList
+                ? [selectedRestrictedList]
+                : this.props.restrictedList;
             deck.status = validateDeck(deck, { packs: this.props.packs, restrictedLists });
         }
 
@@ -94,13 +96,13 @@ class DeckEditor extends React.Component {
     triggerDeckUpdated(restrictedList) {
         const deck = this.getDeckFromState(restrictedList);
 
-        if(this.props.onDeckUpdated) {
+        if (this.props.onDeckUpdated) {
             this.props.onDeckUpdated(deck);
         }
     }
 
     formatCardListItem(card) {
-        if(card.card.custom) {
+        if (card.card.custom) {
             let typeCode = card.card.type;
             let typeName = typeCode[0].toUpperCase() + typeCode.slice(1);
             return card.count + ' Custom ' + typeName + ' - ' + card.card.name;
@@ -135,7 +137,7 @@ class DeckEditor extends React.Component {
             showBanners: this.isAllianceAgenda(selectedAgenda)
         };
 
-        if(!toUpdate.showBanners) {
+        if (!toUpdate.showBanners) {
             toUpdate.bannerCards = [];
         }
 
@@ -149,19 +151,21 @@ class DeckEditor extends React.Component {
     onAddBanner(event) {
         event.preventDefault();
 
-        if(!this.state.selectedBanner || !this.state.selectedBanner.code) {
+        if (!this.state.selectedBanner || !this.state.selectedBanner.code) {
             return;
         }
 
         // Don't allow more than 2 banners
-        if(this.state.bannerCards.length >= 2) {
+        if (this.state.bannerCards.length >= 2) {
             return;
         }
 
         // Don't allow duplicate banners
-        if(this.state.bannerCards.some(banner => {
-            return banner.code === this.state.selectedBanner.code;
-        })) {
+        if (
+            this.state.bannerCards.some((banner) => {
+                return banner.code === this.state.selectedBanner.code;
+            })
+        ) {
             return;
         }
 
@@ -172,7 +176,7 @@ class DeckEditor extends React.Component {
     }
 
     onRemoveBanner(banner) {
-        const banners = this.state.bannerCards.filter(card => {
+        const banners = this.state.bannerCards.filter((card) => {
             return card.code !== banner.code;
         });
 
@@ -186,14 +190,14 @@ class DeckEditor extends React.Component {
     onAddCard(event) {
         event.preventDefault();
 
-        if(!this.state.cardToAdd || !this.state.cardToAdd.label) {
+        if (!this.state.cardToAdd || !this.state.cardToAdd.label) {
             return;
         }
 
         let cardList = this.state.cardList;
         cardList += `${this.state.numberToAdd}  ${this.state.cardToAdd.label}\n`;
 
-        if(this.state.cardToAdd.type === 'plot') {
+        if (this.state.cardToAdd.type === 'plot') {
             let plots = this.state.plotCards;
             this.addCard(plots, this.state.cardToAdd, parseInt(this.state.numberToAdd));
             this.setState({ cardList: cardList, plotCards: plots }, this.triggerDeckUpdated);
@@ -208,45 +212,57 @@ class DeckEditor extends React.Component {
         let split = event.target.value.split('\n');
         let { deckName, faction, agenda, bannerCards, plotCards, drawCards } = this.state;
 
-        let headerMark = split.findIndex(line => line.match(/^Packs:/));
-        if(headerMark >= 0) {
+        let headerMark = split.findIndex((line) => line.match(/^Packs:/));
+        if (headerMark >= 0) {
             // ThronesDB-style deck header found
             // extract deck title, faction, agenda, and banners
-            let header = split.slice(0, headerMark).filter(line => line !== '');
+            let header = split.slice(0, headerMark).filter((line) => line !== '');
             split = split.slice(headerMark);
 
-            if(header.length >= 2) {
+            if (header.length >= 2) {
                 deckName = header[0];
 
-                let newFaction = Object.values(this.props.factions).find(faction => faction.name === header[1].trim());
-                if(newFaction) {
+                let newFaction = Object.values(this.props.factions).find(
+                    (faction) => faction.name === header[1].trim()
+                );
+                if (newFaction) {
                     faction = newFaction;
                 }
 
                 header = header.slice(2);
-                if(header.length >= 1) {
+                if (header.length >= 1) {
                     let rawAgenda, rawBanners;
 
-                    if(header.some(line => {
-                        return line.trim() === 'Alliance';
-                    })) {
+                    if (
+                        header.some((line) => {
+                            return line.trim() === 'Alliance';
+                        })
+                    ) {
                         rawAgenda = 'Alliance';
-                        rawBanners = header.filter(line => line.trim() !== 'Alliance');
+                        rawBanners = header.filter((line) => line.trim() !== 'Alliance');
                     } else {
                         rawAgenda = header[0].trim();
                     }
 
-                    let newAgenda = lookupCardByName({ cardName: rawAgenda, cards: Object.values(this.props.cards), packs: this.props.packs });
-                    if(newAgenda) {
+                    let newAgenda = lookupCardByName({
+                        cardName: rawAgenda,
+                        cards: Object.values(this.props.cards),
+                        packs: this.props.packs
+                    });
+                    if (newAgenda) {
                         agenda = newAgenda;
                     }
 
-                    if(rawBanners) {
+                    if (rawBanners) {
                         let banners = [];
-                        for(let rawBanner of rawBanners) {
-                            let banner = lookupCardByName({ cardName: rawBanner, cards: Object.values(this.props.cards), packs: this.props.packs });
+                        for (let rawBanner of rawBanners) {
+                            let banner = lookupCardByName({
+                                cardName: rawBanner,
+                                cards: Object.values(this.props.cards),
+                                packs: this.props.packs
+                            });
 
-                            if(banner) {
+                            if (banner) {
                                 banners.push(banner);
                             }
                         }
@@ -260,41 +276,48 @@ class DeckEditor extends React.Component {
         plotCards = [];
         drawCards = [];
 
-        for(const line of split) {
+        for (const line of split) {
             let { card, count } = this.parseCardLine(line);
-            if(card) {
+            if (card) {
                 this.addCard(card.type === 'plot' ? plotCards : drawCards, card, count);
             }
         }
 
-        this.setState({
-            cardList: event.target.value,
-            deckName: deckName,
-            faction: faction,
-            agenda: agenda,
-            bannerCards: bannerCards,
-            showBanners: this.isAllianceAgenda(agenda),
-            plotCards: plotCards,
-            drawCards: drawCards
-        }, this.triggerDeckUpdated);
+        this.setState(
+            {
+                cardList: event.target.value,
+                deckName: deckName,
+                faction: faction,
+                agenda: agenda,
+                bannerCards: bannerCards,
+                showBanners: this.isAllianceAgenda(agenda),
+                plotCards: plotCards,
+                drawCards: drawCards
+            },
+            this.triggerDeckUpdated
+        );
     }
 
     parseCardLine(line) {
         const pattern = /^(\d+)x?\s+(.+)$/;
 
         let match = line.trim().match(pattern);
-        if(!match) {
+        if (!match) {
             return { count: 0 };
         }
 
         let count = parseInt(match[1]);
-        let card = lookupCardByName({ cardName: match[2], cards: Object.values(this.props.cards), packs: this.props.packs });
+        let card = lookupCardByName({
+            cardName: match[2],
+            cards: Object.values(this.props.cards),
+            packs: this.props.packs
+        });
 
         return { count: count, card: card };
     }
 
     addCard(list, card, number) {
-        if(list[card.code]) {
+        if (list[card.code]) {
             list[card.code].count += number;
         } else {
             list.push({ count: number, card: card });
@@ -304,21 +327,29 @@ class DeckEditor extends React.Component {
     onSaveClick(event) {
         event.preventDefault();
 
-        if(this.props.onDeckSave) {
+        if (this.props.onDeckSave) {
             this.props.onDeckSave(this.getDeckFromState());
         }
     }
 
     getBannerList() {
-        if(this.state.bannerCards.length === 0) {
+        if (this.state.bannerCards.length === 0) {
             return null;
         }
 
-        return this.state.bannerCards.map(card => {
-            return (<div key={ card.code }>
-                <span key={ card.code } className='card-link col-sm-10'>{ card.label }</span>
-                <span className='glyphicon glyphicon-remove icon-danger btn col-sm-1' aria-hidden='true' onClick={ this.onRemoveBanner.bind(this, card) } />
-            </div>);
+        return this.state.bannerCards.map((card) => {
+            return (
+                <div key={card.code}>
+                    <span key={card.code} className='card-link col-sm-10'>
+                        {card.label}
+                    </span>
+                    <span
+                        className='glyphicon glyphicon-remove icon-danger btn col-sm-1'
+                        aria-hidden='true'
+                        onClick={this.onRemoveBanner.bind(this, card)}
+                    />
+                </div>
+            );
         });
     }
 
@@ -327,91 +358,210 @@ class DeckEditor extends React.Component {
     }
 
     onEventChange(selectedEvent) {
-        this.setState({ eventId: selectedEvent ? selectedEvent._id : undefined }, this.triggerDeckUpdated);
+        this.setState(
+            { eventId: selectedEvent ? selectedEvent._id : undefined },
+            this.triggerDeckUpdated
+        );
     }
 
     getLockedDecksEvents() {
-        return this.props.events.filter(e => e.lockDecks);
+        return this.props.events.filter((e) => e.lockDecks);
     }
 
     render() {
-        if(!this.props.factions || !this.props.agendas || !this.props.cards || !this.props.restrictedList) {
+        if (
+            !this.props.factions ||
+            !this.props.agendas ||
+            !this.props.cards ||
+            !this.props.restrictedList
+        ) {
             return <div>Please wait while loading from the server...</div>;
         }
 
         let banners = this.getBannerList();
-        const cardsExcludingAgendas = Object.values(this.props.cards).filter(card => !this.props.agendas[card.code]);
+        const cardsExcludingAgendas = Object.values(this.props.cards).filter(
+            (card) => !this.props.agendas[card.code]
+        );
 
         let lockedDecksEvents = this.getLockedDecksEvents();
 
         return (
             <div>
-                <ApiStatus apiState={ this.props.apiState } successMessage='Deck saved successfully.' />
+                <ApiStatus
+                    apiState={this.props.apiState}
+                    successMessage='Deck saved successfully.'
+                />
 
                 <div className='form-group'>
                     <div className='col-xs-12 deck-buttons'>
                         <span className='col-xs-2'>
-                            <button ref='submit' type='submit' className='btn btn-primary' onClick={ this.onSaveClick.bind(this) }>Save { this.props.apiState && this.props.apiState.loading && <span className='spinner button-spinner' /> }</button>
+                            <button
+                                ref='submit'
+                                type='submit'
+                                className='btn btn-primary'
+                                onClick={this.onSaveClick.bind(this)}
+                            >
+                                Save{' '}
+                                {this.props.apiState && this.props.apiState.loading && (
+                                    <span className='spinner button-spinner' />
+                                )}
+                            </button>
                         </span>
-                        <button ref='submit' type='button' className='btn btn-primary' onClick={ this.onCancelClick.bind(this) }>Cancel</button>
+                        <button
+                            ref='submit'
+                            type='button'
+                            className='btn btn-primary'
+                            onClick={this.onCancelClick.bind(this)}
+                        >
+                            Cancel
+                        </button>
                     </div>
                 </div>
 
                 <div className='form-group'>
                     <RestrictedListDropdown
-                        currentRestrictedList={ this.props.currentRestrictedList }
-                        onChange={ (restrictedList) => this.triggerDeckUpdated(restrictedList) }
-                        restrictedLists={ this.props.restrictedList }
-                        setCurrentRestrictedList={ this.props.setCurrentRestrictedList } />
+                        currentRestrictedList={this.props.currentRestrictedList}
+                        onChange={(restrictedList) => this.triggerDeckUpdated(restrictedList)}
+                        restrictedLists={this.props.restrictedList}
+                        setCurrentRestrictedList={this.props.setCurrentRestrictedList}
+                    />
                 </div>
 
-                <h4>Either type the cards manually into the box below, add the cards one by one using the card box and autocomplete or for best results, copy and paste a decklist from <a href='http://thronesdb.com' target='_blank'>Thrones DB</a> into the box below.</h4>
+                <h4>
+                    Either type the cards manually into the box below, add the cards one by one
+                    using the card box and autocomplete or for best results, copy and paste a
+                    decklist from{' '}
+                    <a href='http://thronesdb.com' target='_blank'>
+                        Thrones DB
+                    </a>{' '}
+                    into the box below.
+                </h4>
                 <form className='form form-horizontal'>
-                    <Input name='deckName' label='Deck Name' labelClass='col-sm-3' fieldClass='col-sm-9' placeholder='Deck Name'
-                        type='text' onChange={ this.onChange.bind(this, 'deckName') } value={ this.state.deckName } />
-                    <Select name='faction' label='Faction' labelClass='col-sm-3' fieldClass='col-sm-9' options={ Object.values(this.props.factions) }
-                        onChange={ this.onFactionChange.bind(this) } value={ this.state.faction ? this.state.faction.value : undefined } />
-                    <Select name='agenda' label='Agenda' labelClass='col-sm-3' fieldClass='col-sm-9' options={ Object.values(this.props.agendas) }
-                        onChange={ this.onAgendaChange.bind(this) } value={ this.state.agenda ? this.state.agenda.code : undefined }
-                        valueKey='code' nameKey='label' blankOption={ { label: '- Select -', code: '' } } />
+                    <Input
+                        name='deckName'
+                        label='Deck Name'
+                        labelClass='col-sm-3'
+                        fieldClass='col-sm-9'
+                        placeholder='Deck Name'
+                        type='text'
+                        onChange={this.onChange.bind(this, 'deckName')}
+                        value={this.state.deckName}
+                    />
+                    <Select
+                        name='faction'
+                        label='Faction'
+                        labelClass='col-sm-3'
+                        fieldClass='col-sm-9'
+                        options={Object.values(this.props.factions)}
+                        onChange={this.onFactionChange.bind(this)}
+                        value={this.state.faction ? this.state.faction.value : undefined}
+                    />
+                    <Select
+                        name='agenda'
+                        label='Agenda'
+                        labelClass='col-sm-3'
+                        fieldClass='col-sm-9'
+                        options={Object.values(this.props.agendas)}
+                        onChange={this.onAgendaChange.bind(this)}
+                        value={this.state.agenda ? this.state.agenda.code : undefined}
+                        valueKey='code'
+                        nameKey='label'
+                        blankOption={{ label: '- Select -', code: '' }}
+                    />
 
-                    { this.state.showBanners &&
+                    {this.state.showBanners && (
                         <div>
-                            <Select name='banners' label='Banners' labelClass='col-sm-3' fieldClass='col-sm-9' options={ this.props.banners }
-                                onChange={ this.onBannerChange.bind(this) } value={ this.state.selectedBanner ? this.state.selectedBanner.code : undefined }
-                                valueKey='code' nameKey='label'
-                                blankOption={ { label: '- Select -', code: '' } } button={ { text: 'Add', onClick: this.onAddBanner.bind(this) } } />
-                            <div className='col-sm-9 col-sm-offset-3 banner-list'>
-                                { banners }
-                            </div>
+                            <Select
+                                name='banners'
+                                label='Banners'
+                                labelClass='col-sm-3'
+                                fieldClass='col-sm-9'
+                                options={this.props.banners}
+                                onChange={this.onBannerChange.bind(this)}
+                                value={
+                                    this.state.selectedBanner
+                                        ? this.state.selectedBanner.code
+                                        : undefined
+                                }
+                                valueKey='code'
+                                nameKey='label'
+                                blankOption={{ label: '- Select -', code: '' }}
+                                button={{ text: 'Add', onClick: this.onAddBanner.bind(this) }}
+                            />
+                            <div className='col-sm-9 col-sm-offset-3 banner-list'>{banners}</div>
                         </div>
-                    }
-                    <Typeahead label='Card' labelClass={ 'col-sm-3 col-xs-2' } fieldClass='col-sm-4 col-xs-5' labelKey={ 'label' } options={ cardsExcludingAgendas }
-                        onChange={ this.addCardChange.bind(this) }>
-                        <Input name='numcards' type='text' label='Num' labelClass='col-xs-1 no-x-padding' fieldClass='col-xs-2'
-                            value={ this.state.numberToAdd.toString() } onChange={ this.onNumberToAddChange.bind(this) } noGroup>
+                    )}
+                    <Typeahead
+                        label='Card'
+                        labelClass={'col-sm-3 col-xs-2'}
+                        fieldClass='col-sm-4 col-xs-5'
+                        labelKey={'label'}
+                        options={cardsExcludingAgendas}
+                        onChange={this.addCardChange.bind(this)}
+                    >
+                        <Input
+                            name='numcards'
+                            type='text'
+                            label='Num'
+                            labelClass='col-xs-1 no-x-padding'
+                            fieldClass='col-xs-2'
+                            value={this.state.numberToAdd.toString()}
+                            onChange={this.onNumberToAddChange.bind(this)}
+                            noGroup
+                        >
                             <div className='col-xs-1 no-x-padding'>
                                 <div className='btn-group'>
-                                    <button className='btn btn-primary' onClick={ this.onAddCard.bind(this) }>
+                                    <button
+                                        className='btn btn-primary'
+                                        onClick={this.onAddCard.bind(this)}
+                                    >
                                         Add
                                     </button>
                                 </div>
                             </div>
                         </Input>
                     </Typeahead>
-                    <TextArea label='Cards' labelClass='col-sm-3' fieldClass='col-sm-9' rows='10' value={ this.state.cardList }
-                        onChange={ this.onCardListChange.bind(this) } />
-                    { this.state.eventId && 
-                        <h4>Please be aware: Assigning this deck to an event will mean that you will be unable to modify or delete this deck for the duration of the event!</h4>
-                    }
-                    { lockedDecksEvents.length > 0 && 
-                        <Select name='event' label='Event' labelClass='col-sm-3' fieldClass='col-sm-9' options={ lockedDecksEvents }
-                            onChange={ this.onEventChange.bind(this) } value={ this.state.eventId ? this.state.eventId : undefined } 
-                            valueKey='_id' nameKey='name' blankOption={ { label: '- Select -', name: '', value: undefined } }/>
-                    }
+                    <TextArea
+                        label='Cards'
+                        labelClass='col-sm-3'
+                        fieldClass='col-sm-9'
+                        rows='10'
+                        value={this.state.cardList}
+                        onChange={this.onCardListChange.bind(this)}
+                    />
+                    {this.state.eventId && (
+                        <h4>
+                            Please be aware: Assigning this deck to an event will mean that you will
+                            be unable to modify or delete this deck for the duration of the event!
+                        </h4>
+                    )}
+                    {lockedDecksEvents.length > 0 && (
+                        <Select
+                            name='event'
+                            label='Event'
+                            labelClass='col-sm-3'
+                            fieldClass='col-sm-9'
+                            options={lockedDecksEvents}
+                            onChange={this.onEventChange.bind(this)}
+                            value={this.state.eventId ? this.state.eventId : undefined}
+                            valueKey='_id'
+                            nameKey='name'
+                            blankOption={{ label: '- Select -', name: '', value: undefined }}
+                        />
+                    )}
                     <div className='form-group'>
                         <div className='col-sm-offset-3 col-sm-8'>
-                            <button ref='submit' type='submit' className='btn btn-primary' onClick={ this.onSaveClick.bind(this) }>Save { this.props.apiState && this.props.apiState.loading && <span className='spinner button-spinner' /> }</button>
+                            <button
+                                ref='submit'
+                                type='submit'
+                                className='btn btn-primary'
+                                onClick={this.onSaveClick.bind(this)}
+                            >
+                                Save{' '}
+                                {this.props.apiState && this.props.apiState.loading && (
+                                    <span className='spinner button-spinner' />
+                                )}
+                            </button>
                         </div>
                     </div>
                 </form>
